Started work on boss AI
Written baseplate code for boss, in which they can follow the player and decide whether to use close or long range attacks based on their distance from the player. Also added logic for  different boss phases/stages, and logic for randomly choosing attacks within the phase the boss is in to add variety and unpredictability.